refactored out stuff
