+++ experimental/ragas_experimental/embeddings/__init__.py
@@ -1,3 +1,3 @@
-from ragas_experimental.embedding.base import BaseEmbedding, ragas_embedding
+from ragas_experimental.embeddings.base import BaseEmbedding, ragas_embedding
 
 __all__ = ["ragas_embedding", "BaseEmbedding"]
