+++ experimental/ragas_experimental/metric/base.py
@@ -14,11 +14,11 @@ from ragas_experimental.embeddings.base import BaseEmbedding
 from ragas_experimental.prompt.base import Prompt
 from ragas_experimental.prompt.dynamic_few_shot import DynamicFewShotPrompt
 from ragas_experimental.metric.result import MetricResult
+from ragas_experimental.llms import BaseRagasLLM
 from pydantic import BaseModel
 
 if t.TYPE_CHECKING:
     from ragas_experimental.dataset import Dataset
-    from ragas_experimental.llms import BaseRagasLLM
 
 
 @dataclass
