+++ experimental/tests/unit/test_llm.py
@@ -2,7 +2,7 @@ import pytest
 from unittest.mock import Mock
 from pydantic import BaseModel
 
-from ragas_experimental.llm.llm import RagasLLM, ragas_llm
+from ragas_experimental.llms.base import BaseRagasLLM, llm_factory
 
 
 class LLMResponseModel(BaseModel):
@@ -58,59 +58,53 @@ def mock_async_client():
     return MockClient(is_async=True)
 
 
-def test_ragas_llm_initialization(mock_sync_client, monkeypatch):
-    """Test RagasLLM initialization with different providers."""
+def test_llm_factory_initialization(mock_sync_client, monkeypatch):
+    """Test llm_factory initialization with different providers."""
     # Mock instructor to return our mock instructor
     def mock_from_openai(client):
         return MockInstructor(client)
     
     monkeypatch.setattr('instructor.from_openai', mock_from_openai)
     
-    llm = RagasLLM(
-        provider="openai",
-        model="gpt-4",
-        client=mock_sync_client
+    llm = llm_factory(
+        "openai/gpt-4",
+        mock_sync_client
     )
     
-    assert llm.provider == "openai"
     assert llm.model == "gpt-4"
     assert llm.client is not None
     assert not llm.is_async
 
 
-def test_ragas_llm_async_detection(mock_async_client, monkeypatch):
-    """Test that RagasLLM correctly detects async clients."""
+def test_llm_factory_async_detection(mock_async_client, monkeypatch):
+    """Test that llm_factory correctly detects async clients."""
     # Mock instructor to return our mock instructor  
     def mock_from_openai(client):
         return MockInstructor(client)
     
     monkeypatch.setattr('instructor.from_openai', mock_from_openai)
     
-    llm = RagasLLM(
-        provider="openai",
-        model="gpt-4",
-        client=mock_async_client
+    llm = llm_factory(
+        "openai/gpt-4",
+        mock_async_client
     )
     
     assert llm.is_async
 
 
-def test_ragas_llm_factory_function(mock_sync_client, monkeypatch):
-    """Test the ragas_llm factory function."""
+def test_llm_factory_with_model_args(mock_sync_client, monkeypatch):
+    """Test the llm_factory function with model arguments."""
     def mock_from_openai(client):
         return MockInstructor(client)
     
     monkeypatch.setattr('instructor.from_openai', mock_from_openai)
     
-    llm = ragas_llm(
-        provider="openai",
-        model="gpt-4",
-        client=mock_sync_client,
+    llm = llm_factory(
+        "openai/gpt-4",
+        mock_sync_client,
         temperature=0.7
     )
     
-    assert isinstance(llm, RagasLLM)
-    assert llm.provider == "openai"
     assert llm.model == "gpt-4"
     assert llm.model_args.get("temperature") == 0.7
 
@@ -120,10 +114,9 @@ def test_unsupported_provider():
     mock_client = Mock()
     
     with pytest.raises(ValueError, match="Unsupported provider: unsupported"):
-        RagasLLM(
-            provider="unsupported", 
-            model="test-model",
-            client=mock_client
+        llm_factory(
+            "unsupported/test-model",
+            mock_client
         )
 
 
@@ -134,10 +127,9 @@ def test_sync_llm_generate(mock_sync_client, monkeypatch):
     
     monkeypatch.setattr('instructor.from_openai', mock_from_openai)
     
-    llm = RagasLLM(
-        provider="openai",
-        model="gpt-4", 
-        client=mock_sync_client
+    llm = llm_factory(
+        "openai/gpt-4",
+        mock_sync_client
     )
     
     result = llm.generate("Test prompt", LLMResponseModel)
@@ -154,10 +146,9 @@ async def test_async_llm_agenerate(mock_async_client, monkeypatch):
     
     monkeypatch.setattr('instructor.from_openai', mock_from_openai)
     
-    llm = RagasLLM(
-        provider="openai",
-        model="gpt-4",
-        client=mock_async_client
+    llm = llm_factory(
+        "openai/gpt-4",
+        mock_async_client
     )
     
     result = await llm.agenerate("Test prompt", LLMResponseModel)
@@ -173,10 +164,9 @@ def test_sync_client_agenerate_error(mock_sync_client, monkeypatch):
     
     monkeypatch.setattr('instructor.from_openai', mock_from_openai)
     
-    llm = RagasLLM(
-        provider="openai",
-        model="gpt-4",
-        client=mock_sync_client
+    llm = llm_factory(
+        "openai/gpt-4",
+        mock_sync_client
     )
     
     # Test that agenerate raises TypeError with sync client
@@ -200,8 +190,8 @@ def test_provider_support():
         
         # This should not raise an error
         try:
-            llm = RagasLLM(provider=provider, model="test-model", client=mock_client)
-            assert llm.provider == provider
+            llm = llm_factory(f"{provider}/test-model", mock_client)
+            assert llm.model == "test-model"
         except Exception as e:
             pytest.fail(f"Provider {provider} should be supported but got error: {e}")
 
@@ -219,10 +209,9 @@ def test_llm_model_args_storage(mock_sync_client, monkeypatch):
         "top_p": 0.9
     }
     
-    llm = RagasLLM(
-        provider="openai",
-        model="gpt-4",
-        client=mock_sync_client,
+    llm = llm_factory(
+        "openai/gpt-4",
+        mock_sync_client,
         **model_args
     )
